move waitEvent back to lib/tauri
