+++ src/store/source/showProperties.ts
@@ -3,8 +3,8 @@ import { createAsyncThunk } from "@reduxjs/toolkit"
 import { Id, SourceType } from "@/types"
 import { emit } from "@tauri-apps/api/event"
 import { RootState } from ".."
-import { waitEvent } from "@/tauri"
 import { Source } from "."
+import { waitEvent } from "@/lib/tauri"
 import { SourceReader } from "@/lib/source-reader"
 import logger from "@/logger"
